:truck: Moved `LoadingWindow` to `core`
