+++ base/src/main/java/com/arnyminerz/escalaralcoiaicomtat/activity/LoadingActivity.kt
@@ -1,39 +1,13 @@
 package com.arnyminerz.escalaralcoiaicomtat.activity
 
-import android.content.Intent
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
-import androidx.appcompat.content.res.AppCompatResources
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.core.FastOutLinearInEasing
-import androidx.compose.animation.core.animateIntAsState
-import androidx.compose.animation.core.tween
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.LinearProgressIndicator
-import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.arnyminerz.escalaralcoiaicomtat.BuildConfig
-import com.arnyminerz.escalaralcoiaicomtat.R
 import com.arnyminerz.escalaralcoiaicomtat.activity.climb.DataClassActivity
 import com.arnyminerz.escalaralcoiaicomtat.activity.model.NetworkAwareComponentActivity
 import com.arnyminerz.escalaralcoiaicomtat.core.dataClassExploreActivity
@@ -41,13 +15,13 @@ import com.arnyminerz.escalaralcoiaicomtat.core.network.base.ConnectivityProvide
 import com.arnyminerz.escalaralcoiaicomtat.core.preferences.PreferencesModule
 import com.arnyminerz.escalaralcoiaicomtat.core.shared.EXTRA_LINK_PATH
 import com.arnyminerz.escalaralcoiaicomtat.core.shared.sharedPreferences
+import com.arnyminerz.escalaralcoiaicomtat.core.ui.element.LoadingWindow
 import com.arnyminerz.escalaralcoiaicomtat.core.ui.theme.AppTheme
 import com.arnyminerz.escalaralcoiaicomtat.core.utils.getExtra
 import com.arnyminerz.escalaralcoiaicomtat.core.utils.launch
 import com.arnyminerz.escalaralcoiaicomtat.ui.viewmodel.LoadingActivityViewModel
 import com.arnyminerz.escalaralcoiaicomtat.ui.viewmodel.loadingActivityViewModel
 import com.arnyminerz.escalaralcoiaicomtat.view.model.LoadingViewModel
-import com.google.accompanist.drawablepainter.rememberDrawablePainter
 import com.google.firebase.analytics.FirebaseAnalytics
 import com.google.firebase.analytics.ktx.analytics
 import com.google.firebase.ktx.Firebase
@@ -149,8 +123,9 @@ class LoadingActivity : NetworkAwareComponentActivity() {
                 Scaffold { padding ->
                     LoadingWindow(
                         padding,
+                        MainActivity::class.java,
                         stringResource(progressMessageResource, progressMessageAttributes),
-                        errorMessageResource != null,
+                        shouldShowErrorMessage = errorMessageResource != null,
                         if (errorMessageResource != null) stringResource(errorMessageResource!!) else "",
                     )
                 }
@@ -173,93 +148,5 @@ class LoadingActivity : NetworkAwareComponentActivity() {
 
         if (this::loadingViewModel.isInitialized)
             loadingViewModel.tryLoading()
-    }
-}
-
-@Composable
-fun LoadingWindow(
-    padding: PaddingValues,
-    progressMessage: String,
-    shouldShowErrorMessage: Boolean = false,
-    errorMessage: String? = null
-) {
-    val context = LocalContext.current
-
-    val bottomPadding: Int by animateIntAsState(
-        targetValue = if (shouldShowErrorMessage && errorMessage != null) 500 else 0,
-        animationSpec = tween(durationMillis = 300, easing = FastOutLinearInEasing)
-    )
-
-    val drawable = AppCompatResources.getDrawable(LocalContext.current, R.mipmap.ic_launcher_round)
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(padding),
-        contentAlignment = Alignment.Center,
-    ) {
-        Box(
-            modifier = Modifier
-                .align(Alignment.Center)
-                .padding(bottom = bottomPadding.dp)
-        ) {
-            Image(
-                painter = rememberDrawablePainter(drawable),
-                contentDescription = "",
-                modifier = Modifier
-                    .size(128.dp)
-                    .align(Alignment.TopCenter)
-            )
-        }
-        AnimatedVisibility(
-            modifier = Modifier
-                .align(Alignment.Center)
-                .fillMaxWidth(.7f),
-            visible = shouldShowErrorMessage && errorMessage != null
-        ) {
-            Column(
-                horizontalAlignment = Alignment.CenterHorizontally,
-            ) {
-                Text(
-                    text = errorMessage ?: "",
-                    textAlign = TextAlign.Center,
-                    overflow = TextOverflow.Visible
-                )
-                if (BuildConfig.DEBUG)
-                    OutlinedButton(
-                        modifier = Modifier.padding(top = 8.dp),
-                        onClick = {
-                            context.launch(MainActivity::class.java) {
-                                addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-                            }
-                        }
-                    ) {
-                        Text(text = "Enter anyway")
-                    }
-            }
-        }
-
-        AnimatedVisibility(
-            visible = !shouldShowErrorMessage,
-            modifier = Modifier
-                .align(Alignment.BottomStart)
-                .fillMaxWidth()
-        ) {
-            Column(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalAlignment = Alignment.End
-            ) {
-                Text(
-                    modifier = Modifier
-                        .padding(bottom = 9.dp, end = 4.dp),
-                    text = progressMessage,
-                    textAlign = TextAlign.End
-                )
-                LinearProgressIndicator(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                )
-            }
-        }
     }
 }
